+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## Features
+- Added view to basket for *Application Status History* (126577047)
+- Added *Product Limits* for an *Installation* (127177411)
+
 # 1.8.0
 2016-08-12
 
 ## Features
-- Update Variable Deposit slider to use better implementation. 
+- Update Variable Deposit slider to use better implementation.
 
 # 1.7.1
 2016-08-12