+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Features
 - Added view to basket for *Application Status History* (126577047)
 - Added *Product Limits* for an *Installation* (127177411)
+- Added validation for product limits (127177411)
 
 # 1.8.0
 2016-08-12